Change code commit source
Make it more testable, readable. Fix sending pushes over and over,
yet, still sends PRs over and over. Tokens does not work here I
guess. Need to research more on that. There has to be a way to
utilize NextTokens to not process the same PR over and over again.